Update pain.md
Corrected issues with average time interval calculation. 
My fault, adjusted denominator was for the SPC interval chart calculation and not the average measurement calculation on the front tile. 
I have also update the SPC measurement interval rules, but not the pain severity SPC yet. 
Thanks for picking up this mistake, proves the value of working as a team!